clean new testing schema
