+++ .github/workflows/on-commit.yml
@@ -39,7 +39,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest]
         python-version: ['3.10', '3.11']
     steps:
       - uses: actions/checkout@v3
@@ -53,7 +53,8 @@ jobs:
       - run: .venv/bin/python -m pip install pycodestyle pytest nbmake
       - run: .venv/bin/python ./setup.py install
       - run: .venv/bin/python ./setup.py test -a "test_multi.py"
-
+      # needs to be updated to include Windows
+      
   Lint:
     runs-on: ubuntu-latest
     steps:
